HSEARCH-3103 Support setting a sort mode for distance sorts on multi-valued fields
