+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/logging/impl/Log.java
@@ -631,4 +631,8 @@ public interface Log extends BasicLogger {
 			value = "Cannot compute the sum of a temporal field. Only min, max, avg and median are supported.")
 	SearchException cannotComputeSumForTemporalField(@Param EventContext context);
 
+	@Message(id = ID_OFFSET_2 + 117,
+			value = "Cannot compute the sum for a distance sort. Only min, max, avg and median are supported.")
+	SearchException cannotComputeSumForDistanceSort(@Param EventContext context);
+
 }
