+++ engine/src/main/java/org/hibernate/search/engine/search/sort/dsl/DistanceSortOptionsStep.java
@@ -12,6 +12,6 @@ package org.hibernate.search.engine.search.sort.dsl;
  * @param <S> The "self" type (the actual exposed type of this step).
  */
 public interface DistanceSortOptionsStep<S extends DistanceSortOptionsStep<?>>
-		extends SortFinalStep, SortThenStep, SortOrderStep<S> {
+		extends SortFinalStep, SortThenStep, SortOrderStep<S>, SortModeStep<S> {
 
 }
