+++ engine/src/main/java/org/hibernate/search/engine/search/sort/spi/DistanceSortBuilder.java
@@ -6,6 +6,10 @@
  */
 package org.hibernate.search.engine.search.sort.spi;
 
+import org.hibernate.search.engine.search.common.SortMode;
+
 public interface DistanceSortBuilder<B> extends SearchSortBuilder<B> {
+
+	void mode(SortMode mode);
 
 }
